Move `PrintCallback` to `print.py`
