+++ python/ray/train/callbacks/__init__.py
@@ -1,6 +1,7 @@
 from ray.train.callbacks.callback import TrainingCallback
 from ray.train.callbacks.logging import (
-    JsonLoggerCallback, MLflowLoggerCallback, TBXLoggerCallback, PrintCallback)
+    JsonLoggerCallback, MLflowLoggerCallback, TBXLoggerCallback)
+from ray.train.callbacks.print import PrintCallback
 
 __all__ = [
     "TrainingCallback", "JsonLoggerCallback", "MLflowLoggerCallback",
