+++ python/ray/train/callbacks/logging.py
@@ -6,7 +6,6 @@ import logging
 import numpy as np
 import json
 from pathlib import Path
-import collections
 
 from ray.util.debug import log_once
 from ray.util.ml_utils.dict import flatten_dict
@@ -18,36 +17,6 @@ from ray.train.constants import (RESULT_FILE_JSON, TRAINING_ITERATION,
 from ray.util.ml_utils.mlflow import MLflowLoggerUtil
 
 logger = logging.getLogger(__name__)
-
-
-class PrintCallback(TrainingCallback):
-    """A callback that prints training results to STDOUT.
-
-    Example:
-        >>> # Handle results from two workers.
-        >>> callback = PrintCallback()
-        >>> results = [{"loss": 2.290156}, {"loss": 2.275099}]
-        >>> callback.handle_result(results)
-        [
-            {
-                "loss": 2.290156
-            },
-            {
-                "loss": 2.275099
-            }
-        ]
-    """
-
-    def handle_result(self, results: List[Dict], **info):
-        """Prints results to STDOUT.
-
-        Args:
-            results (List[Dict]): List of results from the training
-                function. Each value in the list corresponds to the output of
-                the training function from each worker.
-            **info: kwargs dict for forward compatibility.
-        """
-        print(json.dumps(results, indent=4))
 
 
 class TrainingLogdirMixin:
